Allow Class Resources, call `#to_s` for defined resource_name
If using a class as the resource:

  resource Order do
  end

The resource_name will be a class, which causes issues when sorting for
sections. By calling `#to_s` on the `args.first` when first dealing with
the resource we can ensure that we are always working with a string for
the resource_name.